Updated Sprouts w/ example for splining, Points, displaying lines
This version of the overarching class uses a spline example in the setup, although this should later be removed as it begins to work within the code. Currently it is a demonstration, and used for testing. The draw method was updated to draw all Line object, which were included in this update.
Point appearance changed from rects to ellipses because squares are for squares.