Class for Line created
Line is an Object that is created at a single origin point to begin with, then to which points are/can be added later on. Together the points are joined by the curve() function to form one big line that hits each point. addify(Point) adds another point, and the ArrayList points holds all points.