Added line-drawing functionality
Edited mousePressed so that when mouse is clicked it begins to draw a new Line, adding in a new Point with every click until a black pixel is reached, at which point the line terminates and Nodes are added to the beginning and end of the line. The termination at black pixels simulates the later termination at green pixels that will be added once Nodes are fully functional.