Added stuff on linking Nodes
Altered mousePressed() such that the number of connections made by a Node is kept track of, and once that number reaches three the Node can no longer be used as an anchor point for lines. 
All we have left to do for the actual game is detecting intersections, generating new Node(s) along the middle of the line, and playing turns.