Added displayIntersection to help w/ checks
displayIntersection() is a minor addition, but basically if one has a line, then tacks a point on, not only will the line lengthen, but the previous part of the line will also morph. this previous, morphed section of the line is what is displayed here, without the added part, so that it can be checked whether the new section of the line intersects it or not.