Intersection, all bugs fixed, Node addition
There were several minor bugs throughout the code body that were fixed, these do not require description. 
Adding Nodes was finished, with clauses added so that nodes can only be added on parts of the display with significant differences between blue() and green() values, which only occurs in purple in the context of the colors we used. Furthermore, adding a Node within the radius of another Node was prohibited.
Intersection was created, with essentially blanks the screen, draws the new line, then finds all pixels on that line. The list of pixels created is then checked against all other pixels on the screen, and, with exceptions for finishing lines, if a given pixel which the pine would occupy is colored, then an intersection exists and the line is not allowed. Then, the line is divided into the added section and old section(morphed by the addition), and the added section is checked against that in the way described above and in the recent Line commit in order to determine if intersections exist.